Separated SignInPage & NewAccountPage from Home
